Fix adding a note returns inactive update requests
