+++ submit-api/src/submit_api/resources/staff/package.py
@@ -20,7 +20,7 @@ from flask_restx import Namespace, Resource, cors
 from submit_api.auth import auth
 from submit_api.resources.apihelper import Api as ApiHelper
 from submit_api.schemas.package import CreateUpdateRequestSchema, PackageUpdateRequestSchema, StaffPackageSchema, \
-    CreateUpdateRequestNoteSchema, PackageVersionSchema
+    PackageVersionSchema
 from submit_api.services.package import PackageService
 from submit_api.utils.roles import EpicSubmitRole
 from submit_api.utils.util import cors_preflight
@@ -36,10 +36,6 @@ package_model = ApiHelper.convert_ma_schema_to_restx_model(
 
 create_update_request_model = ApiHelper.convert_ma_schema_to_restx_model(
     API, CreateUpdateRequestSchema(), "CreateUpdateRequest"
-)
-
-create_update_request_note_model = ApiHelper.convert_ma_schema_to_restx_model(
-    API, CreateUpdateRequestNoteSchema(), "CreateUpdateRequestNote"
 )
 
 update_request_model = ApiHelper.convert_ma_schema_to_restx_model(
@@ -129,29 +125,3 @@ class PackageUpdateRequest(Resource):
         """Accept an update request."""
         accept_update_request = PackageService.accept_update_request(package_id)
         return StaffPackageSchema().dump(accept_update_request), HTTPStatus.OK
-
-
-@cors_preflight("POST, OPTIONS")
-@API.route(
-    "/<int:package_id>/update-requests/<int:update_request_id>/note",
-    methods=["POST", "OPTIONS"],
-)
-class PackageUpdateRequestNote(Resource):
-    """Resource for managing a package's update request's note."""
-
-    @staticmethod
-    @ApiHelper.swagger_decorators(API, endpoint_description="Create an update request note for a package")
-    @API.expect(create_update_request_note_model)
-    @API.response(
-        code=HTTPStatus.CREATED, model=package_model, description="Create Update Request Note"
-    )
-    @API.response(HTTPStatus.BAD_REQUEST, "Bad Request")
-    @API.response(HTTPStatus.NOT_FOUND, "Not Found")
-    @auth.require
-    @cors.crossdomain(origin="*")
-    def post(package_id, update_request_id):
-        """Create an update request note."""
-        create_update_request_data = CreateUpdateRequestNoteSchema().load(API.payload)
-        package_with_update_request_note = PackageService.create_update_request_note(
-            package_id, update_request_id, create_update_request_data)
-        return StaffPackageSchema().dump(package_with_update_request_note), HTTPStatus.CREATED
